refactor(admin): split user credit operator services
